Some File cleaning
