+++ RepairsApi/V2/Filtering/ISortConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace RepairsApi.V2.Filtering
+{
+    internal interface ISortConfig<TSearch, TQuery> : ISortItem<TQuery>
+    {
+        bool IsValid(TSearch searchParameter);
+    }
+}
