+++ RepairsApi/V2/Filtering/ISortItem.cs
@@ -1,0 +1,9 @@
+﻿using System.Linq;
+
+namespace RepairsApi.V2.Filtering
+{
+    public interface ISortItem<TQuery>
+    {
+        IQueryable<TQuery> AddOrdering(IQueryable<TQuery> filtered);
+    }
+}
